bug fixes
An error encountered indicates that there was an attempt to access an index out of range in one of its slices. This error probably occurred because the i index in the Train function was used to access both inputs and targets directly, causing a problem when the number of iterations is greater than the number of input examples.